backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
beautifulsoup4==4.12.2
lxml==5.3.0
gunicorn==21.2.0
psycopg>=3.1.0
psycopg-binary>=3.1.0
PyJWT==2.10.1
